api1/requirements.txt
fastapi[all]==0.99.1
pytest==7.4.3
pymongo[srv]==3.11.0
python-dotenv==1.0.0
pydantic==1.10.13
# CORSMiddleware
uvicorn
requests